Replace submissions with target_setting_runs [WIP]
