--- reference_submissions/criteo1tb/tuning_search_space.json
@@ -1,6 +1,0 @@
-{
-  "learning_rate": {"feasible_points": [7e-4]},
-  "warmup_steps": {"feasible_points": [3200]},
-  "weight_decay": {"feasible_points": [1e-4]},
-  "beta1": {"feasible_points": [0.9]}
-}
