--- reference_submissions/imagenet_vit/tuning_search_space.json
@@ -1,9 +1,0 @@
-{
-    "learning_rate": {"feasible_points": [1e-3]},
-    "beta1": {"feasible_points": [0.9]},
-    "beta2": {"feasible_points": [0.999]},
-    "epsilon": {"feasible_points": [1e-8]},
-    "num_epochs": {"feasible_points": [100]},
-    "warmup_epochs": {"feasible_points": [5]},
-    "l2": {"feasible_points": [1e-1]}
-}
