--- reference_submissions/librispeech/tuning_search_space.json
@@ -1,4 +1,0 @@
-{
-    "learning_rate": {"feasible_points": [0.0001, 0.0002, 0.0003]}
-}
-
