--- reference_submissions/mnist/tuning_search_space.json
@@ -1,5 +1,0 @@
-{
-    "learning_rate": {"min": 1e-4, "max": 1e-2, "scaling": "log"},
-    "one_minus_beta_1": {"min": 0.9, "max": 0.999, "scaling": "log"},
-    "epsilon": {"feasible_points": [1e-8, 1e-5, 1e-3]}
-}
