--- reference_submissions/ogbg/ogbg_jax/submission.py
@@ -1,121 +1,0 @@
-from typing import Dict, Iterator, List, Tuple
-
-from flax import jax_utils
-import jax
-from jax import lax
-import jax.numpy as jnp
-import optax
-
-from algorithmic_efficiency import spec
-
-
-def get_batch_size(workload_name):
-  # Return the global batch size.
-  batch_sizes = {'ogbg': 2048}
-  return batch_sizes[workload_name]
-
-
-def init_optimizer_state(workload: spec.Workload,
-                         model_params: spec.ParameterContainer,
-                         model_state: spec.ModelAuxiliaryState,
-                         hyperparameters: spec.Hyperparameters,
-                         rng: spec.RandomState) -> spec.OptimizerState:
-  """Creates an Adam optimizer."""
-  del model_params
-  del model_state
-  del rng
-  params_zeros_like = jax.tree_map(lambda s: jnp.zeros(s.shape_tuple),
-                                   workload.param_shapes)
-  opt_init_fn, opt_update_fn = opt_init_fn, opt_update_fn = optax.adam(
-      learning_rate=hyperparameters.learning_rate)
-  optimizer_state = opt_init_fn(params_zeros_like)
-  return jax_utils.replicate(optimizer_state), opt_update_fn
-
-
-def train_step(workload,
-               opt_update_fn,
-               model_state,
-               optimizer_state,
-               current_param_container,
-               hyperparameters,
-               batch,
-               rng):
-  del hyperparameters
-
-  def loss_fn(params):
-    logits_batch, new_model_state  = workload.model_fn(
-        params,
-        batch,
-        model_state,
-        spec.ForwardPassMode.TRAIN,
-        rng,
-        update_batch_norm=True)
-    mask_batch = batch['weights']
-    per_example_losses = workload.loss_fn(batch['targets'],
-                                          logits_batch,
-                                          mask_batch)
-    mean_loss = (
-        jnp.sum(jnp.where(mask_batch, per_example_losses, 0)) /
-        jnp.sum(mask_batch))
-    return mean_loss, new_model_state
-
-  grad_fn = jax.value_and_grad(loss_fn, has_aux=True)
-  (_, new_model_state), grad = grad_fn(current_param_container)
-  grad = lax.pmean(grad, axis_name='batch')
-  updates, new_optimizer_state = opt_update_fn(
-      grad, optimizer_state, current_param_container)
-  updated_params = optax.apply_updates(current_param_container, updates)
-  return new_model_state, new_optimizer_state, updated_params
-
-
-def update_params(
-    workload: spec.Workload,
-    current_param_container: spec.ParameterContainer,
-    current_params_types: spec.ParameterTypeTree,
-    model_state: spec.ModelAuxiliaryState,
-    hyperparameters: spec.Hyperparameters,
-    batch: Dict[str, spec.Tensor],
-    loss_type: spec.LossType,
-    # This will define the output activation via `output_activation_fn`.
-    optimizer_state: spec.OptimizerState,
-    eval_results: List[Tuple[int, float]],
-    global_step: int,
-    rng: spec.RandomState) -> spec.UpdateReturn:
-  """Return (updated_optimizer_state, updated_params, updated_model_state)."""
-  del current_params_types
-  del loss_type
-  del eval_results
-  del global_step
-
-  optimizer_state, opt_update_fn = optimizer_state
-  pmapped_train_step = jax.pmap(
-      train_step,
-      axis_name='batch',
-      in_axes=(None, None, 0, 0, 0, None, 0, 0),
-      static_broadcasted_argnums=(0, 1))
-  dropout_rngs = jax.random.split(rng, jax.local_device_count())
-  new_model_state, new_optimizer_state, new_params = pmapped_train_step(
-      workload, opt_update_fn, model_state, optimizer_state,
-      current_param_container, hyperparameters, batch, dropout_rngs)
-  return (new_optimizer_state, opt_update_fn), new_params, new_model_state
-
-
-def data_selection(
-    workload: spec.Workload,
-    input_queue: Iterator[Dict[str, spec.Tensor]],
-    optimizer_state: spec.OptimizerState,
-    current_param_container: spec.ParameterContainer,
-    hyperparameters: spec.Hyperparameters,
-    global_step: int,
-    rng: spec.RandomState) -> Tuple[spec.Tensor, spec.Tensor, spec.Tensor]:
-  """Select data from the infinitely repeating, pre-shuffled input queue.
-
-  Each element of the queue is a batch of training examples and labels.
-  """
-  del workload
-  del optimizer_state
-  del current_param_container
-  del hyperparameters
-  del global_step
-  del rng
-  return next(input_queue)
