--- reference_submissions/ogbg/tuning_search_space.json
@@ -1,1 +1,0 @@
-{"learning_rate": {"feasible_points": [1e-3]}}
