+++ target_setting_runs/ogbg/tuning_search_space.json
@@ -1,0 +1,37 @@
+{
+    "learning_rate": {
+        "feasible_points": [
+            0.000487
+        ]
+    },
+    "beta1": {
+        "feasible_points": [
+            0.8194
+        ]
+    },
+    "beta2": {
+        "feasible_points": [
+            0.9803
+        ]
+    },
+    "epsilon": {
+        "feasible_points": [
+            1e-8
+        ]
+    },
+    "warmup_steps": {
+        "feasible_points": [
+            1200
+        ]
+    },
+    "num_steps": {
+        "feasible_points": [
+            60000
+        ]
+    },
+    "l2": {
+        "feasible_points": [
+            0.407336
+        ]
+    }
+}
