+++ target_setting_runs/pytorch_adamw.py
@@ -1,0 +1,108 @@
+"""Submission file for an AdamW optimizer in PyTorch."""
+
+from typing import Dict, Iterator, List, Tuple
+
+import torch
+from torch.optim.lr_scheduler import CosineAnnealingLR
+from torch.optim.lr_scheduler import LinearLR
+from torch.optim.lr_scheduler import SequentialLR
+
+from algorithmic_efficiency import spec
+
+
+def init_optimizer_state(workload: spec.Workload,
+                         model_params: spec.ParameterContainer,
+                         model_state: spec.ModelAuxiliaryState,
+                         hyperparameters: spec.Hyperparameters,
+                         rng: spec.RandomState) -> spec.OptimizerState:
+  """Creates an AdamW optimizer."""
+  del model_state
+  del rng
+
+  optimizer_state = {
+      'optimizer':
+          torch.optim.AdamW(
+              model_params.parameters(),
+              lr=hyperparameters.learning_rate,
+              betas=(hyperparameters.beta1, hyperparameters.beta2),
+              eps=hyperparameters.epsilon,
+              weight_decay=hyperparameters.l2,
+          )
+  }
+
+  scheduler1 = LinearLR(
+      optimizer_state['optimizer'],
+      start_factor=1e-10,
+      end_factor=1.,
+      total_iters=hyperparameters.warmup_steps)
+  cosine_steps = max(hyperparameters.num_steps - hyperparameters.warmup_steps,
+                     1)
+  scheduler2 = CosineAnnealingLR(
+      optimizer_state['optimizer'], T_max=cosine_steps)
+
+  optimizer_state['scheduler'] = SequentialLR(
+      optimizer_state['optimizer'],
+      schedulers=[scheduler1, scheduler2],
+      milestones=[hyperparameters.warmup_steps])
+
+  return optimizer_state
+
+
+def update_params(workload: spec.Workload,
+                  current_param_container: spec.ParameterContainer,
+                  current_params_types: spec.ParameterTypeTree,
+                  model_state: spec.ModelAuxiliaryState,
+                  hyperparameters: spec.Hyperparameters,
+                  batch: Dict[str, spec.Tensor],
+                  loss_type: spec.LossType,
+                  optimizer_state: spec.OptimizerState,
+                  eval_results: List[Tuple[int, float]],
+                  global_step: int,
+                  rng: spec.RandomState) -> spec.UpdateReturn:
+  """Return (updated_optimizer_state, updated_params, updated_model_state)."""
+  del current_params_types
+  del hyperparameters
+  del loss_type
+  del eval_results
+  del global_step
+
+  current_model = current_param_container
+  current_model.train()
+  optimizer_state['optimizer'].zero_grad()
+
+  logits, new_model_state = workload.model_fn(
+      params=current_model,
+      augmented_and_preprocessed_input_batch=batch,
+      model_state=model_state,
+      mode=spec.ForwardPassMode.TRAIN,
+      rng=rng,
+      update_batch_norm=True)
+
+  mask = batch['weights']
+  per_example_losses = workload.loss_fn(batch['targets'], logits, mask)
+  loss = torch.where(mask, per_example_losses, 0).sum() / mask.sum()
+
+  loss.backward()
+  optimizer_state['optimizer'].step()
+
+  return optimizer_state, current_param_container, new_model_state
+
+
+def data_selection(
+    workload: spec.Workload,
+    input_queue: Iterator[Dict[str, spec.Tensor]],
+    optimizer_state: spec.OptimizerState,
+    current_param_container: spec.ParameterContainer,
+    hyperparameters: spec.Hyperparameters,
+    global_step: int,
+    rng: spec.RandomState) -> Tuple[spec.Tensor, spec.Tensor, spec.Tensor]:
+  """Select data from the infinitely repeating, pre-shuffled input queue.
+  Each element of the queue is a batch of training examples and labels.
+  """
+  del workload
+  del optimizer_state
+  del current_param_container
+  del hyperparameters
+  del global_step
+  del rng
+  return next(input_queue)
